Split post 1414 oversized slices
